analyser/requirements.txt
aenum==3.1.11
aiohttp==3.8.3
aiohttp-cors==0.7.0
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
distlib==0.3.4
dnspython==2.3.0
envparse==0.2.0
filelock==3.7.1
frozenlist==1.3.3
gobject==0.1.0
idna==3.3
kaggle==1.5.12
lark==1.1.5
motor==3.1.1
multidict==6.0.3
platformdirs==2.5.2
pycairo==1.21.0
pycodestyle==2.10.0
pycryptodomex==3.16.0
pygccxml==2.2.1
PyGObject==3.42.2
pygtrie==2.5.0
pymongo==4.3.3
python-dateutil==2.8.2
python-ndn==0.3.post2
python-slugify==6.1.2
requests==2.28.1
scapy==2.5.0
six==1.16.0
text-unidecode==1.3
tomli==2.0.1
tqdm==4.64.0
urllib3==1.26.11
yarl==1.8.2
